feat: wait for sender on Close
Waits for 20 seconds on Close for any queued logs to be sent.

Also makes `Send` synchronous so we can guarantee log delivery for a
script that makes a request, calls Close and then immediately exits.
This should be okay in terms of concurrency as we have a buffer in the
FanIn channel, and also the Sender is sending asyncronously and dropping
logs when it has too many to handle (channel throughput should be high)